+++ src/env.js
@@ -15,7 +15,8 @@ const envSchema = z.object({
   WALLET_PK: z.string(),
   AUTHORITY_DID_WEB: z.string().refine(val => /^did:[^:]+:[^:]+$/.test(val), {
     message: "AUTHORITY_DID_WEB must be in the format 'did:{string}:{string}'"
-  })
+  }),
+  STORACHA_LIT_ACTION_CID: z.string().default('QmRGzM3MUHoMMWynEH5BxNTaZsKCeVphYQTiCLNHTyfWzX')
 })
 
 // validate `process.env` against our schema
